Systematic baselines (#7)
* Rewrite how we pass '_reps' on to graphnet and transformer: TF-models get tensorreps as argument for in/hidden/out channels; handle in_reps and out_reps in the same way; check that out_reps are scalar-only; dont inherit config files of TF models from base models (base models are just for debugging, will not be in the paper); update TFTransformer and TFGraphnet documentation

* Fix tests (didnt adapt them in the last commit, oopsie)

* Fixed another stupid mistake: forgot out_reps argument in GraphNetWrapper and TransformerWrapper

* Pick graphnet and transformer model hyperparameters to be more similar to PN and ParT

* Pick graphnet and transformer training hyperparameters to be more similar to PN and ParT

* Optimized learning rate for transformer (tried 1e-3, 3e-4, 1e-4)

* Modify graphnet to match ParticleNet more closely: 3 linear layers per edge_conv, reduce the hidden_reps to keep the model small (250k params)

* Small modification to logging: Print lframesnet #params; print in_reps and out_reps in tagging

* Print only first 7 chars of git hash (as commonly displayed on github)

* Shortcut tensorreps_transforms also if we only have scalar reps

* Add comment that #channels should stay fixed in graphnet and transformer

* Boilerplate rewrite for orthogonalization: Add use_float64 as option; code style improvements (dont access kwargs that might not be defined)

+++ tensorframes/lframes/equi_lframes.py
@@ -75,7 +75,6 @@ class OrthogonalLearnedLFrames(LearnedLFrames):
 
     def forward(self, fourmomenta, scalars, edge_index, batch, return_tracker=False):
         vecs = super().forward(fourmomenta, scalars, edge_index)
-        vecs = vecs.to(dtype=torch.float64)
         vecs = [vecs[..., i, :] for i in range(self.n_vectors)]
 
         trafo, reg_lightlike, reg_coplanar = orthogonal_trafo(
@@ -83,7 +82,7 @@ class OrthogonalLearnedLFrames(LearnedLFrames):
         )
 
         tracker = {"reg_lightlike": reg_lightlike, "reg_coplanar": reg_coplanar}
-        lframes = LFrames(trafo.to(dtype=scalars.dtype))
+        lframes = LFrames(trafo)
         return (lframes, tracker) if return_tracker else lframes
 
 
@@ -107,9 +106,7 @@ class RestLFrames(LearnedLFrames):
 
     def forward(self, fourmomenta, scalars, edge_index, batch, return_tracker=False):
         references = super().forward(fourmomenta, scalars, edge_index)
-        references = references.to(dtype=torch.float64)
         references = [references[..., i, :] for i in range(self.n_vectors)]
-        fourmomenta = fourmomenta.to(dtype=torch.float64)
 
         trafo, reg_collinear = restframe_equivariant(
             fourmomenta,
@@ -118,7 +115,7 @@ class RestLFrames(LearnedLFrames):
             return_reg=True,
         )
         tracker = {"reg_collinear": reg_collinear}
-        lframes = LFrames(trafo.to(dtype=scalars.dtype))
+        lframes = LFrames(trafo)
         return (lframes, tracker) if return_tracker else lframes
 
 
@@ -144,7 +141,6 @@ class LearnedRestLFrames(LearnedLFrames):
 
     def forward(self, fourmomenta, scalars, edge_index, batch, return_tracker=False):
         vecs = super().forward(fourmomenta, scalars, edge_index)
-        vecs = vecs.to(dtype=torch.float64)
         fourmomenta = vecs[..., 0, :]
         references = [vecs[..., i, :] for i in range(1, self.n_vectors)]
 
@@ -155,5 +151,5 @@ class LearnedRestLFrames(LearnedLFrames):
             return_reg=True,
         )
         tracker = {"reg_collinear": reg_collinear}
-        lframes = LFrames(trafo.to(dtype=scalars.dtype))
+        lframes = LFrames(trafo)
         return (lframes, tracker) if return_tracker else lframes
